Move load profile dir to components and network items to network subdir
